change package name to corp standarts
